fix(breadcrumb): relocate methods to meet 500-line ceiling
- Relocate ShowPopup and PublishPopupMessengerReady from
  BreadcrumbDropDownHost.cs into BreadcrumbDropDownHost.Open.cs,
  bringing the first file from 514 back under the repo's 500-line
  ceiling (498 lines) with zero behavior change
- Add composition regression test for #680/#677 focus-predicate interaction
- Correct two stale statements in the delivery report via an
  append-only addendum

Refs: #680
