+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/other/p3-t1-cr1-addendum-verified.2026-08-28T19-32.md
@@ -1,0 +1,9 @@
+Timestamp: 2026-08-28T19-32
+Command: git diff HEAD -- docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/delivery-report.2026-08-28T16-40.md | Select-String -Pattern '^-[^-]'; Select-String -SimpleMatch "Post-Rebase Addendum" <file>; Select-String -SimpleMatch "Correction 1: the scheduled action calls FocusPending(), not the raw _focusPending delegate." <file>; Select-String -SimpleMatch "Correction 2: issue #677 has since merged into this branch's base and the shipped code composes with its MayTakeFocus machinery." <file>
+EXIT_CODE: 0
+Output Summary:
+- Deletion-pattern matches (^-[^-], anchored to HEAD): 0 (edit is append-only, no existing line deleted or altered)
+- "Post-Rebase Addendum" hits: 1
+- "Correction 1: the scheduled action calls FocusPending(), not the raw _focusPending delegate." hits: 1
+- "Correction 2: issue #677 has since merged into this branch's base and the shipped code composes with its MayTakeFocus machinery." hits: 1
+All acceptance conditions satisfied.
