+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/qa-gates/p4-t2-analyzers.2026-08-28T19-42.md
@@ -1,0 +1,7 @@
+Timestamp: 2026-08-28T19-42
+Command: MSBuild.exe TaskMaster.sln -t:Rebuild -m -p:Configuration=Debug -p:Platform="Any CPU" -p:EnableNETAnalyzers=true -p:EnforceCodeStyleInBuild=true
+EXIT_CODE: 0
+Output Summary: Build succeeded. 5 Warning(s), 0 Error(s). Warning set matches BASELINE_ANALYZER_WARNINGS
+(P0-T4) exactly: identical System.Reactive.PackagesConfigCheck warnings across the same 5 projects
+(QuickFiler.csproj, TaskMaster.csproj, ToDoModel.csproj, UtilitiesCS.Test.csproj, UtilitiesCS.csproj).
+0 new diagnostics.
