+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/qa-gates/p4-t3-nullable.2026-08-28T19-46.md
@@ -1,0 +1,5 @@
+Timestamp: 2026-08-28T19-46
+Command: MSBuild.exe TaskMaster.sln -t:Rebuild -m -p:Configuration=Debug -p:Platform="Any CPU" -p:TreatWarningsAsErrors=true
+EXIT_CODE: 0
+Output Summary: Build succeeded. 5 Warning(s), 0 Error(s). Zero CS86xx nullable diagnostics. Warning set
+matches BASELINE_NULLABLE_WARNINGS (P0-T5) exactly (same System.Reactive.PackagesConfigCheck set).
