+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/qa-gates/p4-t4-qft-full-run.2026-08-28T19-50.md
@@ -1,0 +1,5 @@
+Timestamp: 2026-08-28T19-50
+Command: vstest.console.exe QuickFiler.Test/bin/Debug/QuickFiler.Test.dll /Settings:scripts/vscode/TaskMaster.cli.runsettings /InIsolation /TestCaseFilter:"TestCategory!=LiveOutlook" /Logger:"trx;LogFileName=p4-t4.trx" /ResultsDirectory:<FEATURE>/evidence/qa-gates/p4-t4
+EXIT_CODE: 0
+Output Summary: Test Run Successful. Total: 1237 (BASELINE_QFT_COUNT = 1236 from P0-T7 + 1 new CR-2
+test). Passed: 1237. Failed: 0. Results directory holds exactly one file: p4-t4.trx.
