+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/qa-gates/p4-t6-file-size-audit.2026-08-28T20-10.md
@@ -1,0 +1,7 @@
+Timestamp: 2026-08-28T20-10
+Command: (Get-Content QuickFiler\Viewers\BreadcrumbDropDownHost.cs).Count ; (Get-Content QuickFiler\Viewers\BreadcrumbDropDownHost.Open.cs).Count
+EXIT_CODE: 0
+Output Summary: BreadcrumbDropDownHost.cs = 498 lines (<= 500; strictly less than BASELINE_HOST_COUNT
+= 514). BreadcrumbDropDownHost.Open.cs = 107 lines (<= 500; strictly greater than BASELINE_OPEN_COUNT
+= 90). The R1 file-size-ceiling violation is closed, not deferred: both files are within the
+repository's 500-line ceiling.
