+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/qa-gates/p4-t7-commit-readiness.2026-08-28T20-12.md
@@ -1,0 +1,10 @@
+Timestamp: 2026-08-28T20-12
+Command: git status --porcelain
+EXIT_CODE: 0
+Output Summary: All porcelain entries fall into one of the allowed categories: the two relocation-touched
+production files (QuickFiler/Viewers/BreadcrumbDropDownHost.cs, QuickFiler/Viewers/BreadcrumbDropDownHost.Open.cs),
+the one test file (QuickFiler.Test/Viewers/BreadcrumbDropDownHostTests.Part3.cs), the delivery report
+(delivery-report.2026-08-28T16-40.md), this remediation plan file itself, or a path under
+docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/.
+Programmatic check over every listed path: count of entries matching none of these categories = 0.
+No `git commit` performed by this task.
