+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/regression-testing/p1-t3-hosttests-post-move.2026-08-28T19-09.md
@@ -1,0 +1,6 @@
+Timestamp: 2026-08-28T19-09
+Command: vstest.console.exe QuickFiler.Test/bin/Debug/QuickFiler.Test.dll /Settings:scripts/vscode/TaskMaster.cli.runsettings /InIsolation /TestCaseFilter:"FullyQualifiedName~BreadcrumbDropDownHostTests" /Logger:"trx;LogFileName=p1-t3.trx" /ResultsDirectory:<FEATURE>/evidence/regression-testing/p1-t3
+EXIT_CODE: 0
+Output Summary: Test Run Successful. Total: 35 (matches BASELINE_HOSTTESTS_COUNT = 35 from P0-T6
+exactly). Passed: 35. Failed: 0. Zero test deltas from the relocation.
+Results directory holds exactly one file: p1-t3.trx.
