+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/regression-testing/p2-t1-test-added.2026-08-28T19-15.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-08-28T19-15
+Command: Select-String -SimpleMatch "OpenAsync_TakeFocusReopenAfterNonCapturingOpenWithPredicateFalse_RestoresAutoCloseButSuppressesFocus" QuickFiler.Test\Viewers\BreadcrumbDropDownHostTests.Part3.cs; (Get-Content QuickFiler.Test\Viewers\BreadcrumbDropDownHostTests.Part3.cs).Count
+EXIT_CODE: 0
+Output Summary: New test name present exactly 1 time. File line count is 425 (<= 500 ceiling).
