+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/regression-testing/p2-t2-build.2026-08-28T19-17.md
@@ -1,0 +1,5 @@
+Timestamp: 2026-08-28T19-17
+Command: MSBuild.exe TaskMaster.sln -t:Build -m -p:Configuration=Debug -p:Platform="Any CPU"
+EXIT_CODE: 0
+Output Summary: Build succeeded. 5 Warning(s) (same pre-existing System.Reactive PackagesConfigCheck
+set), 0 Error(s). The new test compiles cleanly.
