+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/regression-testing/p2-t3-new-test-green.2026-08-28T19-27.md
@@ -1,0 +1,12 @@
+Timestamp: 2026-08-28T19-27
+Command: vstest.console.exe QuickFiler.Test/bin/Debug/QuickFiler.Test.dll /Settings:scripts/vscode/TaskMaster.cli.runsettings /InIsolation /TestCaseFilter:"FullyQualifiedName~BreadcrumbDropDownHostTests" /Logger:"trx;LogFileName=p2-t3.trx" /ResultsDirectory:<FEATURE>/evidence/regression-testing/p2-t3
+EXIT_CODE: 0
+Output Summary: Test Run Successful. Total: 36 (BASELINE_HOSTTESTS_COUNT = 35 + 1 new test). Passed: 36.
+Failed: 0. New test OpenAsync_TakeFocusReopenAfterNonCapturingOpenWithPredicateFalse_RestoresAutoCloseButSuppressesFocus
+appears in the TRX with outcome="Passed" (12 ms).
+Results directory holds exactly one file: p2-t3.trx.
+
+Observation (not a plan task, recorded for the final report): the raw TRX carries vstest's own
+`computerName` attribute (a local machine name), which this plan does not budget a sanitization task
+for. D6 in this plan scopes host-path hygiene to the `Command:`/`Output Summary:` fields composed in
+this and other evidence artifacts, which contain no host-identifying text.
