+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/p0-t2-context.2026-08-28T18-16.md
@@ -1,0 +1,22 @@
+Timestamp: 2026-08-28T18-16
+Command: git rev-parse HEAD; git status --porcelain; (Get-Content QuickFiler\Viewers\BreadcrumbDropDownHost.cs).Count; (Get-Content QuickFiler\Viewers\BreadcrumbDropDownHost.Open.cs).Count
+EXIT_CODE: 0
+Output Summary:
+- REMEDIATION_BASE_COMMIT (informational only, not asserted against a fixed hash): 8e82a2e07cbef741b36dc18045c1fec685b13842
+- git status --porcelain observed output (NOT empty; see deviation note below):
+  ` M docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/remediation-plan.2026-08-28T17-15.md`
+  `?? docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/`
+- BASELINE_HOST_COUNT: 514 (matches D2's expected ~514)
+- BASELINE_OPEN_COUNT: 90 (matches D2's expected ~90)
+
+Deviation note (recorded, not concealed): the plan's P0-T2 acceptance condition states
+"git status --porcelain output is empty at the moment this task runs." At the moment P0-T2 actually
+ran, the working tree was NOT empty of changes, but the only two entries present are this same
+remediation session's own P0-T1 output, produced per the atomic-executor check-off protocol before
+P0-T2 began: (1) the `[ ]` -> `[x]` check-off edit to this remediation plan file for P0-T1, and
+(2) the newly created `evidence/remediation-baseline/` directory holding P0-T1's own artifact. Prior
+to P0-T1's edits, `git status --porcelain` was confirmed empty (recorded separately, immediately before
+Phase 0 began). No pre-existing outstanding review/memory artifact is present; the orchestrator's
+precondition (a clean tree before Phase 0 starts) was satisfied. This deviation is escalated in the
+executor's final completion report rather than treated as a Phase 0 blocker, per the executor's
+blocking protocol (blocking is permitted only pre-execution).
