+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/p0-t3-format.2026-08-28T18-18.md
@@ -1,0 +1,4 @@
+Timestamp: 2026-08-28T18-18
+Command: dotnet tool run csharpier check .
+EXIT_CODE: 0
+Output Summary: Checked 1560 files in 4189ms. No formatting drift; baseline is clean.
