+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/p0-t5-nullable.2026-08-28T18-25.md
@@ -1,0 +1,6 @@
+Timestamp: 2026-08-28T18-25
+Command: MSBuild.exe TaskMaster.sln -t:Rebuild -m -p:Configuration=Debug -p:Platform="Any CPU" -p:TreatWarningsAsErrors=true
+EXIT_CODE: 0
+Output Summary: Build succeeded. 5 Warning(s), 0 Error(s). Zero CS86xx nullable diagnostics.
+BASELINE_NULLABLE_WARNINGS = 5, identical System.Reactive.PackagesConfigCheck warning set as P0-T4:
+QuickFiler.csproj, TaskMaster.csproj, ToDoModel.csproj, UtilitiesCS.Test.csproj, UtilitiesCS.csproj.
