+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/p0-t6-hosttests-baseline.2026-08-28T18-30.md
@@ -1,0 +1,5 @@
+Timestamp: 2026-08-28T18-30
+Command: vstest.console.exe QuickFiler.Test/bin/Debug/QuickFiler.Test.dll /Settings:scripts/vscode/TaskMaster.cli.runsettings /InIsolation /TestCaseFilter:"FullyQualifiedName~BreadcrumbDropDownHostTests" /Logger:"trx;LogFileName=p0-t6.trx" /ResultsDirectory:<FEATURE>/evidence/remediation-baseline/p0-t6
+EXIT_CODE: 0
+Output Summary: Test Run Successful. BASELINE_HOSTTESTS_COUNT = 35. Passed: 35. Failed: 0.
+Results directory holds exactly one file: p0-t6.trx.
