+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/p0-t7-qft-baseline.2026-08-28T18-32.md
@@ -1,0 +1,5 @@
+Timestamp: 2026-08-28T18-32
+Command: vstest.console.exe QuickFiler.Test/bin/Debug/QuickFiler.Test.dll /Settings:scripts/vscode/TaskMaster.cli.runsettings /InIsolation /TestCaseFilter:"TestCategory!=LiveOutlook" /Logger:"trx;LogFileName=p0-t7.trx" /ResultsDirectory:<FEATURE>/evidence/remediation-baseline/p0-t7
+EXIT_CODE: 0
+Output Summary: Test Run Successful. BASELINE_QFT_COUNT = 1236. Passed: 1236. Failed: 0.
+Results directory holds exactly one file: p0-t7.trx.
