+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/p0-t8-coverage-baseline.2026-08-28T18-55.md
@@ -1,0 +1,13 @@
+Timestamp: 2026-08-28T18-55
+Command: pwsh -NoProfile -File ./scripts/vscode/Invoke-MSTestWithCoverage.ps1 -SearchRoot . -CoverageOutput coverage/coverage-remediation-baseline-680.cobertura.xml (run detached per D4, polled to completion)
+EXIT_CODE: 0 (inferred from the run's own success markers: "Test Run Successful", 0 failures reported,
+cobertura artifact successfully written and post-processed for Koverage compatibility, no error text in
+the run log; the process was launched detached, so no direct shell exit-code capture was available —
+consistent with D4's note that the specific polling/exit-observation mechanism is an execution detail)
+Output Summary:
+- BASELINE_COVERAGE_TOTAL = 6856; Passed: 6856; Failed: 0
+- BASELINE_FAILURE_SET = none
+- BASELINE_LINE_RATE = 0.852841 (85.2841%)
+- BASELINE_BRANCH_RATE = 0.79234 (79.234%)
+- Coverage artifact: coverage/coverage-remediation-baseline-680.cobertura.xml (root <coverage> element:
+  lines-covered=54749, lines-valid=64196, branches-covered=13034, branches-valid=16450)
