+++ docs/features/active/2026-08-28-quickfiler-search-box-loses-focus-on-dropdown-expand-680/evidence/remediation-baseline/phase0-instructions-read.md
@@ -1,0 +1,17 @@
+Timestamp: 2026-08-28T18-15
+
+Policy Order:
+1. CLAUDE.md
+2. .claude/rules/general-code-change.md
+3. .claude/rules/general-unit-test.md
+4. .claude/rules/csharp.md
+5. .claude/rules/quality-tiers.md
+6. .claude/rules/tonality.md
+
+Files read (explicit list, all six, in the order above):
+- CLAUDE.md
+- .claude/rules/general-code-change.md
+- .claude/rules/general-unit-test.md
+- .claude/rules/csharp.md
+- .claude/rules/quality-tiers.md
+- .claude/rules/tonality.md
